+++ services/bls_aggregation/blsagg_test.go
@@ -1007,7 +1007,7 @@ func TestBlsAgg(t *testing.T) {
 			taskResponseDigest2, err := hashFunction(taskResponse2)
 			require.Nil(t, err)
 			blsSigOp2 := testOperator2BlsKeypair.SignMessage(taskResponseDigest2)
-			ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+			ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
 			defer cancel()
 			taskSignature2 := NewTaskSignature(taskIndex, taskResponse2, blsSigOp2, testOperator2OperatorId)
 			err = blsAggServ.ProcessNewSignature(